settle

+++ hyunsub-dutch/src/main/kotlin/kim/hyunsub/dutch/model/dto/DutchRecordMemberCreateParams.kt
@@ -2,6 +2,6 @@ package kim.hyunsub.dutch.model.dto
 
 data class DutchRecordMemberCreateParams(
 	val memberId: String,
-	val actual: Int,
-	val should: Int,
+	val actual: Double,
+	val should: Double,
 )
